14. niepubliczna klasa node
